extract TypeHeaderShape
